+++ src/compas_model/interactions/__init__.py
@@ -1,0 +1,5 @@
+from .interaction import Interaction
+
+__all__ = [
+    "Interaction",
+]
